bandit_mask creates class method for defined bits
This means calling bandit_mask in a class defines a class level method
which delegates to BanditMask.bits, so, for example, when working with a
form in Rails, you can refer to User.roles instead of RoleMask.bits.

Side note: I found a bug in the way bandit_mask checked whether or not
it would be overwriting existing instance methoods. The checks were
wrong, but the tests were also wrong, so it wasn't being caught. This
should be fixed now.