bit names are always symbols, but allow strings
Because this change has the potential to result in lots of symbols being
created, I'm bumping the minimum ruby version. 2.2+ includes symbol GC
which should mitigate the risk of runaway symbol creation.

This change will eliminate the need to jump through hoops in web
applications that take user input (as strings) that would otherwise need
to be converted to symbols in the application before using the mask.